Reorganizing the parsers code
