+++ redis/asyncio/__init__.py
@@ -7,7 +7,6 @@ from redis.asyncio.connection import (
     SSLConnection,
     UnixDomainSocketConnection,
 )
-from redis.asyncio.parser import CommandsParser
 from redis.asyncio.sentinel import (
     Sentinel,
     SentinelConnectionPool,
@@ -38,7 +37,6 @@ __all__ = [
     "BlockingConnectionPool",
     "BusyLoadingError",
     "ChildDeadlockedError",
-    "CommandsParser",
     "Connection",
     "ConnectionError",
     "ConnectionPool",
