+++ redis/commands/__init__.py
@@ -1,7 +1,6 @@
 from .cluster import READ_COMMANDS, AsyncRedisClusterCommands, RedisClusterCommands
 from .core import AsyncCoreCommands, CoreCommands
 from .helpers import list_or_args
-from .parser import CommandsParser
 from .redismodules import AsyncRedisModuleCommands, RedisModuleCommands
 from .sentinel import AsyncSentinelCommands, SentinelCommands
 
@@ -10,7 +9,6 @@ __all__ = [
     "AsyncRedisClusterCommands",
     "AsyncRedisModuleCommands",
     "AsyncSentinelCommands",
-    "CommandsParser",
     "CoreCommands",
     "READ_COMMANDS",
     "RedisClusterCommands",
