+++ redis/connection.py
@@ -1,512 +1,48 @@
 import copy
-import errno
-import io
 import os
 import socket
+import ssl
 import threading
 import weakref
-from io import SEEK_END
 from itertools import chain
 from queue import Empty, Full, LifoQueue
 from time import time
-from typing import Optional, Union
+from typing import Optional, Type, Union
 from urllib.parse import parse_qs, unquote, urlparse
 
-from redis.backoff import NoBackoff
-from redis.credentials import CredentialProvider, UsernamePasswordCredentialProvider
-from redis.exceptions import (
+from .backoff import NoBackoff
+from .credentials import CredentialProvider, UsernamePasswordCredentialProvider
+from .exceptions import (
     AuthenticationError,
     AuthenticationWrongNumberOfArgsError,
-    BusyLoadingError,
     ChildDeadlockedError,
     ConnectionError,
     DataError,
-    ExecAbortError,
-    InvalidResponse,
-    ModuleError,
-    NoPermissionError,
-    NoScriptError,
-    ReadOnlyError,
     RedisError,
     ResponseError,
     TimeoutError,
 )
-from redis.retry import Retry
-from redis.utils import CRYPTOGRAPHY_AVAILABLE, HIREDIS_AVAILABLE, str_if_bytes
-
-try:
-    import ssl
-
-    ssl_available = True
-except ImportError:
-    ssl_available = False
-
-NONBLOCKING_EXCEPTION_ERROR_NUMBERS = {BlockingIOError: errno.EWOULDBLOCK}
-
-if ssl_available:
-    if hasattr(ssl, "SSLWantReadError"):
-        NONBLOCKING_EXCEPTION_ERROR_NUMBERS[ssl.SSLWantReadError] = 2
-        NONBLOCKING_EXCEPTION_ERROR_NUMBERS[ssl.SSLWantWriteError] = 2
-    else:
-        NONBLOCKING_EXCEPTION_ERROR_NUMBERS[ssl.SSLError] = 2
-
-NONBLOCKING_EXCEPTIONS = tuple(NONBLOCKING_EXCEPTION_ERROR_NUMBERS.keys())
-
-if HIREDIS_AVAILABLE:
-    import hiredis
+from .parsers import Encoder, _HiredisParser, _RESP2Parser
+from .retry import Retry
+from .utils import (
+    CRYPTOGRAPHY_AVAILABLE,
+    HIREDIS_AVAILABLE,
+    SSL_AVAILABLE,
+    str_if_bytes,
+)
 
 SYM_STAR = b"*"
 SYM_DOLLAR = b"$"
 SYM_CRLF = b"\r\n"
 SYM_EMPTY = b""
 
-SERVER_CLOSED_CONNECTION_ERROR = "Connection closed by server."
-
 SENTINEL = object()
-MODULE_LOAD_ERROR = "Error loading the extension. Please check the server logs."
-NO_SUCH_MODULE_ERROR = "Error unloading module: no such module with that name"
-MODULE_UNLOAD_NOT_POSSIBLE_ERROR = "Error unloading module: operation not possible."
-MODULE_EXPORTS_DATA_TYPES_ERROR = (
-    "Error unloading module: the module "
-    "exports one or more module-side data "
-    "types, can't unload"
-)
-# user send an AUTH cmd to a server without authorization configured
-NO_AUTH_SET_ERROR = {
-    # Redis >= 6.0
-    "AUTH <password> called without any password "
-    "configured for the default user. Are you sure "
-    "your configuration is correct?": AuthenticationError,
-    # Redis < 6.0
-    "Client sent AUTH, but no password is set": AuthenticationError,
-}
 
-
-class Encoder:
-    "Encode strings to bytes-like and decode bytes-like to strings"
-
-    def __init__(self, encoding, encoding_errors, decode_responses):
-        self.encoding = encoding
-        self.encoding_errors = encoding_errors
-        self.decode_responses = decode_responses
-
-    def encode(self, value):
-        "Return a bytestring or bytes-like representation of the value"
-        if isinstance(value, (bytes, memoryview)):
-            return value
-        elif isinstance(value, bool):
-            # special case bool since it is a subclass of int
-            raise DataError(
-                "Invalid input of type: 'bool'. Convert to a "
-                "bytes, string, int or float first."
-            )
-        elif isinstance(value, (int, float)):
-            value = repr(value).encode()
-        elif not isinstance(value, str):
-            # a value we don't know how to deal with. throw an error
-            typename = type(value).__name__
-            raise DataError(
-                f"Invalid input of type: '{typename}'. "
-                f"Convert to a bytes, string, int or float first."
-            )
-        if isinstance(value, str):
-            value = value.encode(self.encoding, self.encoding_errors)
-        return value
-
-    def decode(self, value, force=False):
-        "Return a unicode string from the bytes-like representation"
-        if self.decode_responses or force:
-            if isinstance(value, memoryview):
-                value = value.tobytes()
-            if isinstance(value, bytes):
-                value = value.decode(self.encoding, self.encoding_errors)
-        return value
-
-
-class BaseParser:
-    EXCEPTION_CLASSES = {
-        "ERR": {
-            "max number of clients reached": ConnectionError,
-            "invalid password": AuthenticationError,
-            # some Redis server versions report invalid command syntax
-            # in lowercase
-            "wrong number of arguments "
-            "for 'auth' command": AuthenticationWrongNumberOfArgsError,
-            # some Redis server versions report invalid command syntax
-            # in uppercase
-            "wrong number of arguments "
-            "for 'AUTH' command": AuthenticationWrongNumberOfArgsError,
-            MODULE_LOAD_ERROR: ModuleError,
-            MODULE_EXPORTS_DATA_TYPES_ERROR: ModuleError,
-            NO_SUCH_MODULE_ERROR: ModuleError,
-            MODULE_UNLOAD_NOT_POSSIBLE_ERROR: ModuleError,
-            **NO_AUTH_SET_ERROR,
-        },
-        "WRONGPASS": AuthenticationError,
-        "EXECABORT": ExecAbortError,
-        "LOADING": BusyLoadingError,
-        "NOSCRIPT": NoScriptError,
-        "READONLY": ReadOnlyError,
-        "NOAUTH": AuthenticationError,
-        "NOPERM": NoPermissionError,
-    }
-
-    def parse_error(self, response):
-        "Parse an error response"
-        error_code = response.split(" ")[0]
-        if error_code in self.EXCEPTION_CLASSES:
-            response = response[len(error_code) + 1 :]
-            exception_class = self.EXCEPTION_CLASSES[error_code]
-            if isinstance(exception_class, dict):
-                exception_class = exception_class.get(response, ResponseError)
-            return exception_class(response)
-        return ResponseError(response)
-
-
-class SocketBuffer:
-    def __init__(
-        self, socket: socket.socket, socket_read_size: int, socket_timeout: float
-    ):
-        self._sock = socket
-        self.socket_read_size = socket_read_size
-        self.socket_timeout = socket_timeout
-        self._buffer = io.BytesIO()
-
-    def unread_bytes(self) -> int:
-        """
-        Remaining unread length of buffer
-        """
-        pos = self._buffer.tell()
-        end = self._buffer.seek(0, SEEK_END)
-        self._buffer.seek(pos)
-        return end - pos
-
-    def _read_from_socket(
-        self,
-        length: Optional[int] = None,
-        timeout: Union[float, object] = SENTINEL,
-        raise_on_timeout: Optional[bool] = True,
-    ) -> bool:
-        sock = self._sock
-        socket_read_size = self.socket_read_size
-        marker = 0
-        custom_timeout = timeout is not SENTINEL
-
-        buf = self._buffer
-        current_pos = buf.tell()
-        buf.seek(0, SEEK_END)
-        if custom_timeout:
-            sock.settimeout(timeout)
-        try:
-            while True:
-                data = self._sock.recv(socket_read_size)
-                # an empty string indicates the server shutdown the socket
-                if isinstance(data, bytes) and len(data) == 0:
-                    raise ConnectionError(SERVER_CLOSED_CONNECTION_ERROR)
-                buf.write(data)
-                data_length = len(data)
-                marker += data_length
-
-                if length is not None and length > marker:
-                    continue
-                return True
-        except socket.timeout:
-            if raise_on_timeout:
-                raise TimeoutError("Timeout reading from socket")
-            return False
-        except NONBLOCKING_EXCEPTIONS as ex:
-            # if we're in nonblocking mode and the recv raises a
-            # blocking error, simply return False indicating that
-            # there's no data to be read. otherwise raise the
-            # original exception.
-            allowed = NONBLOCKING_EXCEPTION_ERROR_NUMBERS.get(ex.__class__, -1)
-            if not raise_on_timeout and ex.errno == allowed:
-                return False
-            raise ConnectionError(f"Error while reading from socket: {ex.args}")
-        finally:
-            buf.seek(current_pos)
-            if custom_timeout:
-                sock.settimeout(self.socket_timeout)
-
-    def can_read(self, timeout: float) -> bool:
-        return bool(self.unread_bytes()) or self._read_from_socket(
-            timeout=timeout, raise_on_timeout=False
-        )
-
-    def read(self, length: int) -> bytes:
-        length = length + 2  # make sure to read the \r\n terminator
-        # BufferIO will return less than requested if buffer is short
-        data = self._buffer.read(length)
-        missing = length - len(data)
-        if missing:
-            # fill up the buffer and read the remainder
-            self._read_from_socket(missing)
-            data += self._buffer.read(missing)
-        return data[:-2]
-
-    def readline(self) -> bytes:
-        buf = self._buffer
-        data = buf.readline()
-        while not data.endswith(SYM_CRLF):
-            # there's more data in the socket that we need
-            self._read_from_socket()
-            data += buf.readline()
-
-        return data[:-2]
-
-    def get_pos(self) -> int:
-        """
-        Get current read position
-        """
-        return self._buffer.tell()
-
-    def rewind(self, pos: int) -> None:
-        """
-        Rewind the buffer to a specific position, to re-start reading
-        """
-        self._buffer.seek(pos)
-
-    def purge(self) -> None:
-        """
-        After a successful read, purge the read part of buffer
-        """
-        unread = self.unread_bytes()
-
-        # Only if we have read all of the buffer do we truncate, to
-        # reduce the amount of memory thrashing.  This heuristic
-        # can be changed or removed later.
-        if unread > 0:
-            return
-
-        if unread > 0:
-            # move unread data to the front
-            view = self._buffer.getbuffer()
-            view[:unread] = view[-unread:]
-        self._buffer.truncate(unread)
-        self._buffer.seek(0)
-
-    def close(self) -> None:
-        try:
-            self._buffer.close()
-        except Exception:
-            # issue #633 suggests the purge/close somehow raised a
-            # BadFileDescriptor error. Perhaps the client ran out of
-            # memory or something else? It's probably OK to ignore
-            # any error being raised from purge/close since we're
-            # removing the reference to the instance below.
-            pass
-        self._buffer = None
-        self._sock = None
-
-
-class PythonParser(BaseParser):
-    "Plain Python parsing class"
-
-    def __init__(self, socket_read_size):
-        self.socket_read_size = socket_read_size
-        self.encoder = None
-        self._sock = None
-        self._buffer = None
-
-    def __del__(self):
-        try:
-            self.on_disconnect()
-        except Exception:
-            pass
-
-    def on_connect(self, connection):
-        "Called when the socket connects"
-        self._sock = connection._sock
-        self._buffer = SocketBuffer(
-            self._sock, self.socket_read_size, connection.socket_timeout
-        )
-        self.encoder = connection.encoder
-
-    def on_disconnect(self):
-        "Called when the socket disconnects"
-        self._sock = None
-        if self._buffer is not None:
-            self._buffer.close()
-            self._buffer = None
-        self.encoder = None
-
-    def can_read(self, timeout):
-        return self._buffer and self._buffer.can_read(timeout)
-
-    def read_response(self, disable_decoding=False):
-        pos = self._buffer.get_pos()
-        try:
-            result = self._read_response(disable_decoding=disable_decoding)
-        except BaseException:
-            self._buffer.rewind(pos)
-            raise
-        else:
-            self._buffer.purge()
-            return result
-
-    def _read_response(self, disable_decoding=False):
-        raw = self._buffer.readline()
-        if not raw:
-            raise ConnectionError(SERVER_CLOSED_CONNECTION_ERROR)
-
-        byte, response = raw[:1], raw[1:]
-
-        if byte not in (b"-", b"+", b":", b"$", b"*"):
-            raise InvalidResponse(f"Protocol Error: {raw!r}")
-
-        # server returned an error
-        if byte == b"-":
-            response = response.decode("utf-8", errors="replace")
-            error = self.parse_error(response)
-            # if the error is a ConnectionError, raise immediately so the user
-            # is notified
-            if isinstance(error, ConnectionError):
-                raise error
-            # otherwise, we're dealing with a ResponseError that might belong
-            # inside a pipeline response. the connection's read_response()
-            # and/or the pipeline's execute() will raise this error if
-            # necessary, so just return the exception instance here.
-            return error
-        # single value
-        elif byte == b"+":
-            pass
-        # int value
-        elif byte == b":":
-            response = int(response)
-        # bulk response
-        elif byte == b"$":
-            length = int(response)
-            if length == -1:
-                return None
-            response = self._buffer.read(length)
-        # multi-bulk response
-        elif byte == b"*":
-            length = int(response)
-            if length == -1:
-                return None
-            response = [
-                self._read_response(disable_decoding=disable_decoding)
-                for i in range(length)
-            ]
-        if isinstance(response, bytes) and disable_decoding is False:
-            response = self.encoder.decode(response)
-        return response
-
-
-class HiredisParser(BaseParser):
-    "Parser class for connections using Hiredis"
-
-    def __init__(self, socket_read_size):
-        if not HIREDIS_AVAILABLE:
-            raise RedisError("Hiredis is not installed")
-        self.socket_read_size = socket_read_size
-        self._buffer = bytearray(socket_read_size)
-
-    def __del__(self):
-        try:
-            self.on_disconnect()
-        except Exception:
-            pass
-
-    def on_connect(self, connection, **kwargs):
-        self._sock = connection._sock
-        self._socket_timeout = connection.socket_timeout
-        kwargs = {
-            "protocolError": InvalidResponse,
-            "replyError": self.parse_error,
-            "errors": connection.encoder.encoding_errors,
-        }
-
-        if connection.encoder.decode_responses:
-            kwargs["encoding"] = connection.encoder.encoding
-        self._reader = hiredis.Reader(**kwargs)
-        self._next_response = False
-
-    def on_disconnect(self):
-        self._sock = None
-        self._reader = None
-        self._next_response = False
-
-    def can_read(self, timeout):
-        if not self._reader:
-            raise ConnectionError(SERVER_CLOSED_CONNECTION_ERROR)
-
-        if self._next_response is False:
-            self._next_response = self._reader.gets()
-            if self._next_response is False:
-                return self.read_from_socket(timeout=timeout, raise_on_timeout=False)
-        return True
-
-    def read_from_socket(self, timeout=SENTINEL, raise_on_timeout=True):
-        sock = self._sock
-        custom_timeout = timeout is not SENTINEL
-        try:
-            if custom_timeout:
-                sock.settimeout(timeout)
-            bufflen = self._sock.recv_into(self._buffer)
-            if bufflen == 0:
-                raise ConnectionError(SERVER_CLOSED_CONNECTION_ERROR)
-            self._reader.feed(self._buffer, 0, bufflen)
-            # data was read from the socket and added to the buffer.
-            # return True to indicate that data was read.
-            return True
-        except socket.timeout:
-            if raise_on_timeout:
-                raise TimeoutError("Timeout reading from socket")
-            return False
-        except NONBLOCKING_EXCEPTIONS as ex:
-            # if we're in nonblocking mode and the recv raises a
-            # blocking error, simply return False indicating that
-            # there's no data to be read. otherwise raise the
-            # original exception.
-            allowed = NONBLOCKING_EXCEPTION_ERROR_NUMBERS.get(ex.__class__, -1)
-            if not raise_on_timeout and ex.errno == allowed:
-                return False
-            raise ConnectionError(f"Error while reading from socket: {ex.args}")
-        finally:
-            if custom_timeout:
-                sock.settimeout(self._socket_timeout)
-
-    def read_response(self, disable_decoding=False):
-        if not self._reader:
-            raise ConnectionError(SERVER_CLOSED_CONNECTION_ERROR)
-
-        # _next_response might be cached from a can_read() call
-        if self._next_response is not False:
-            response = self._next_response
-            self._next_response = False
-            return response
-
-        if disable_decoding:
-            response = self._reader.gets(False)
-        else:
-            response = self._reader.gets()
-
-        while response is False:
-            self.read_from_socket()
-            if disable_decoding:
-                response = self._reader.gets(False)
-            else:
-                response = self._reader.gets()
-        # if the response is a ConnectionError or the response is a list and
-        # the first item is a ConnectionError, raise it as something bad
-        # happened
-        if isinstance(response, ConnectionError):
-            raise response
-        elif (
-            isinstance(response, list)
-            and response
-            and isinstance(response[0], ConnectionError)
-        ):
-            raise response[0]
-        return response
-
-
-DefaultParser: BaseParser
+DefaultParser: Type[Union[_RESP2Parser, _HiredisParser]]
 if HIREDIS_AVAILABLE:
-    DefaultParser = HiredisParser
+    DefaultParser = _HiredisParser
 else:
-    DefaultParser = PythonParser
+    DefaultParser = _RESP2Parser
 
 
 class Connection:
@@ -987,7 +523,7 @@ class SSLConnection(Connection):
         Raises:
             RedisError
         """  # noqa
-        if not ssl_available:
+        if not SSL_AVAILABLE:
             raise RedisError("Python wasn't built with SSL support")
 
         super().__init__(**kwargs)
