+++ redis/typing.py
@@ -1,14 +1,23 @@
 # from __future__ import annotations
 
 from datetime import datetime, timedelta
-from typing import TYPE_CHECKING, Any, Awaitable, Iterable, TypeVar, Union
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Awaitable,
+    Iterable,
+    Mapping,
+    Type,
+    TypeVar,
+    Union,
+)
 
 from redis.compat import Protocol
 
 if TYPE_CHECKING:
     from redis.asyncio.connection import ConnectionPool as AsyncConnectionPool
-    from redis.asyncio.connection import Encoder as AsyncEncoder
-    from redis.connection import ConnectionPool, Encoder
+    from redis.connection import ConnectionPool
+    from redis.parsers import Encoder
 
 
 Number = Union[int, float]
@@ -39,6 +48,8 @@ AnyKeyT = TypeVar("AnyKeyT", bytes, str, memoryview)
 AnyFieldT = TypeVar("AnyFieldT", bytes, str, memoryview)
 AnyChannelT = TypeVar("AnyChannelT", bytes, str, memoryview)
 
+ExceptionMappingT = Mapping[str, Union[Type[Exception], Mapping[str, Type[Exception]]]]
+
 
 class CommandsProtocol(Protocol):
     connection_pool: Union["AsyncConnectionPool", "ConnectionPool"]
@@ -48,7 +59,7 @@ class CommandsProtocol(Protocol):
 
 
 class ClusterCommandsProtocol(CommandsProtocol):
-    encoder: Union["AsyncEncoder", "Encoder"]
+    encoder: Encoder
 
     def execute_command(self, *args, **options) -> Union[Any, Awaitable]:
         ...
