+++ whitelist.py
@@ -14,6 +14,5 @@ exc_type  # unused variable (/data/repos/redis/redis-py/redis/asyncio/utils.py:2
 exc_value  # unused variable (/data/repos/redis/redis-py/redis/asyncio/utils.py:26)
 traceback  # unused variable (/data/repos/redis/redis-py/redis/asyncio/utils.py:26)
 AsyncConnectionPool  # unused import (//data/repos/redis/redis-py/redis/typing.py:9)
-AsyncEncoder  # unused import (//data/repos/redis/redis-py/redis/typing.py:10)
 AsyncRedis  # unused import (//data/repos/redis/redis-py/redis/commands/core.py:49)
 TargetNodesT  # unused import (//data/repos/redis/redis-py/redis/commands/cluster.py:46)
